Test schedule bugs fixed
