+++ candidate/templates/candidate-index.html
@@ -38,7 +38,7 @@
                     <li id="edi" ><a href="{% url 'candidate:edit' pk=pk %}"><span class="fa fa-pencil"></span>Edit Profile</a></li>
                     <li id="app"><a href="{% url 'candidate:applications' pk=pk %}"><span class="fa fa-file-text-o"></span>Applications</a></li>
                     <li id="fav"><a href="{% url 'candidate:favjobs' pk=pk %}"><span class="fa fa-heart-o"></span>Favourite Jobs</a></li>
-                    <li id="cha" ><a href="{% url 'candidate:change' pk=pk %}"><span class="fa fa-lock"></span>Change Password</a></li>
+                    {% comment %}<li id="cha" ><a href="{% url 'candidate:change' pk=pk %}"><span class="fa fa-lock"></span>Change Password</a></li>{% endcomment %}
                     <li id="tes" ><a href="{% url 'candidate:tests' pk=pk %}"><span class="fa fa-question"></span>Tests</a></li>
                     <li id="ins" ><a href="{% url 'candidate:interviews' pk=pk %}"><span class="fa fa-user"></span>Interviews</a></li>
                     {% comment %} <li><a href="candidate-dashboard-applications.html"><span class="fa fa-file-text-o"></span>Apllications</a></li>
@@ -140,7 +140,7 @@
                                     <li class="nav-item"><a href="{% url 'candidate:edit' pk=pk %}"><span class="fa fa-pencil"></span>Edit Profile</a></li>
                                     <li class="nav-item"><a href="{% url 'candidate:applications' pk=pk %}"><span class="fa fa-file-text-o"></span>Applications</a></li>
                                     <li class="nav-item"><a href="{% url 'candidate:favjobs' pk=pk %}"><span class="fa fa-heart-o"></span>Favourite Jobs</a></li>
-                                    <li class="nav-item"><a href="{% url 'candidate:change' pk=pk %}"><span class="fa fa-lock"></span>Change Password</a></li>
+                                    {% comment %} <li class="nav-item"><a href="{% url 'candidate:change' pk=pk %}"><span class="fa fa-lock"></span>Change Password</a></li>{% endcomment %}
                                     <li class="nav-item"><a href="{% url 'candidate:tests' pk=pk %}"><span class="fa fa-question"></span>Tests</a></li>
                                     <li class="nav-item"><a href="{% url 'candidate:interviews' pk=pk %}"><span class="fa fa-user"></span>Interviews</a></li>
                                     {% comment %} <li class="nav-item"><a href="candidate-dashboard-applications.html"><span class="fa fa-file-text-o"></span>Apllications</a></li>
