+++ candidate/templates/dashboard-candidate.html
@@ -100,10 +100,11 @@
     </div>
 
     <div class="row mt-4 mt-lg-5">
-        {% for i in notifics %}
         <div class="col-xxl-6">
             <h2>Recent Notifications</h2>
             <div class="pxp-candidate-dashboard-notifications">
+        {% for i in notifics %}
+
                 {% if i.notif_type == "M" %}
                 <div class="pxp-candidate-dashboard-notifications-item mb-3">
                     <div class="pxp-candidate-dashboard-notifications-item-left">
@@ -145,12 +146,14 @@
                     <div class="pxp-candidate-dashboard-notifications-item-right">{{ i.datetime|naturaltime }}</div>
                 </div>
                 {% endif %}
+                {% endfor%}
+
             </div>
         </div>
-                {% endfor%}
-                {% for i in recent %}
                 <div class="col-xxl-6">
                     <h2 class="mt-4 mt-lg-5 mt-xxl-0">Recent Messages</h2>
+                {% for i in recent %}
+
                     <div class="pxp-candidate-dashboard-messages">
                 <div class="pxp-candidate-dashboard-messages-item mb-3">
                     {% if i.3 %}
@@ -165,15 +168,17 @@
                     </div>
                 </div>
             </div>
+            {% endfor %}
+
         </div>
-                {% endfor %}
     </div>
 
     <div class="mt-4 mt-lg-5">
-        {% for i in applications %}
         <h2>Recent Job Applications</h2>
         <div class="table-responsive">
             <table class="table align-middle">
+                {% for i in applications %}
+
                 <tr>
                     {% if i.logo %}
                         <td style="width: 3%;"><div class="pxp-candidate-dashboard-job-avatar pxp-cover"  style="background-image: url('{{ i.logo.url }}')"></div></td>
@@ -195,9 +200,10 @@
                         </div>
                     </td>
                 </tr>
+                {% endfor %}
+
             </table>
         </div>
-                {% endfor %}
     </div>
 </div>
 <div class="modal fade pxp-user-modal" id="pxp-signup-modal" aria-hidden="true" aria-labelledby="signupModal" tabindex="-1">
