+++ candidate/urls.py
@@ -42,7 +42,7 @@ urlpatterns = [
 	path('thread/', views.thread, name='thread'),
 	path('company/', views.company, name='company'),
 	path('tests/', views.tests, name='tests'),
-	path('attempt/<int:pk2>/', views.attempt, name='attempt'),
+	path('attempt/<int:pk2>/<int:pk3>/', views.attempt, name='attempt'),
 	path('submit/', views.submit, name='submit'),
     path('interviews/', views.interviews, name='interviews'),
     path('get_interview/', views.get_interview, name='get_interview'),
