+++ employer/templates/newjob-employer.html
@@ -102,6 +102,15 @@
                     </select>
                 </div>
             </div>
+            <div class="col-md-6 col-xxl-3">
+                <div class="mb-3">
+                    <label for="pxp-company-job-salary" title = "Request candidates also from mpower admin"class="form-label">Request candidates</label>
+                    <select id="pxp-company-job-salary" class="form-select" name="reqCand">
+                        <option value="1">No</option>
+                        <option value="2">Yes</option>
+                    </select>
+                </div>
+            </div>
         </div>
 
         <div class="mt-4 mt-lg-5">
