+++ employer/templates/suggestions-employer.html
@@ -123,7 +123,7 @@
                                     {% if GET_params %}
                                     <li class="page-item"><a style="padding: 0.7rem" href="?{{GET_params.urlencode}}&amp;page={{ i }}" class="page-link">{{i}}</a></li>
                                     {% else %}
-                                    <li class="page-item"><a style="padding: 0.7rem" href="?page={{ i }}" class="page-link">{{i}}</a></li>
+                                    <li class="page-item"><a style="padding: 0.7rem" href="?job_id={{ sel }}&amp;page={{ i }}" class="page-link">{{i}}</a></li>
                                     {% endif %}
                                 {% endif %}
                             {% endfor %}
@@ -132,7 +132,7 @@
                                 {% if GET_params %}
                                 <li class="page-item"><a style="padding: 0.7rem" href="?{{GET_params.urlencode}}&amp;page={{ pe.number|add:'5' }}" class="page-link">&hellip;</a></li>
                                 {% else %}
-                                <li class="page-item"><a style="padding: 0.7rem" href="?page={{ pe.number|add:'5' }}" class="page-link">&hellip;</a></li>
+                                <li class="page-item"><a style="padding: 0.7rem" href="?page={{ pe.number|add:'5' }}&job_id={{sel}}" class="page-link">&hellip;</a></li>
                                 {% endif %}
                             {% endif %}
 
@@ -140,7 +140,7 @@
                                 {% if GET_params %}
                                 <li class="page-item"><a style="padding: 0.7rem" href="?{{GET_params.urlencode}}&amp;page={{ pe.next_page_number }}" class="page-link"><i class="fa fa-chevron-right" aria-hidden="true"></i></a></li>
                                 {% else %}
-                                <li class="page-item"><a style="padding: 0.7rem" href="?page={{pe.next_page_number}}" class="page-link"><i class="fa fa-chevron-right" aria-hidden="true"></i></a></li>
+                                <li class="page-item"><a style="padding: 0.7rem" href="?page={{pe.next_page_number}}&job_id={{sel}}" class="page-link"><i class="fa fa-chevron-right" aria-hidden="true"></i></a></li>
                                 {% endif %}
                             {% endif %}
                             {% endif %}
