+++ employer/urls.py
@@ -42,5 +42,8 @@ urlpatterns = [
     path('done/', views.done, name='done'),
     path('get_link/', views.get_link, name='get_link'),
 	path('logout/', views.logout, name='clogout'),
+	path('suggest/', views.suggest, name='suggest'),
+	path('request_cand/', views.requestCand, name='requestCand'),
+	path('interview_complete/', views.interview_complete, name='interview_complete'),
 	path('under_development/', views.under_development, name='under_development'),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
