+++ main/models.py
@@ -271,20 +271,7 @@ class TestQues(models.Model):
         db_table = "testques"
 
 
-class TestUser(models.Model):
-    testuser_id = models.AutoField(primary_key=True)
-    user_id = models.ForeignKey(
-        JobSeeker, on_delete=models.CASCADE, blank=True, null=True)
-    test_id = models.ForeignKey(
-        Test, on_delete=models.CASCADE, blank=True, null=True)
-    correct_answers = models.IntegerField(default=0)
-    total_ques = models.IntegerField(default=0)
-    answers = models.CharField(max_length=100, default=None, null=True)
-    date = models.DateTimeField(default=timezone.now)
-    
 
-    class Meta:
-        db_table = "testuser"
 
 
 class Application(models.Model):
@@ -305,6 +292,25 @@ class Application(models.Model):
     class Meta:
         db_table = "application"
 
+class TestUser(models.Model):
+    testuser_id = models.AutoField(primary_key=True)
+    user_id = models.ForeignKey(
+        JobSeeker, on_delete=models.CASCADE, blank=True, null=True)
+    test_id = models.ForeignKey(
+        Test, on_delete=models.CASCADE, blank=True, null=True)
+    emp_id = models.ForeignKey(
+        Employer, on_delete=models.CASCADE, blank=True, null=True)
+    correct_answers = models.IntegerField(default=0)
+    total_ques = models.IntegerField(default=0)
+    answers = models.CharField(max_length=100, default=None, null=True)
+    date = models.DateTimeField(default=timezone.now)
+    apply_id = models.ForeignKey(
+        Application, on_delete=models.CASCADE, blank=True, null=True)
+    
+    
+
+    class Meta:
+        db_table = "testuser"
 
 class Notifications(models.Model):
     notif_id = models.AutoField(primary_key=True)
@@ -330,6 +336,8 @@ class Interview(models.Model):
         JobSeeker, on_delete=models.CASCADE, blank=True, null=True)
     eid = models.ForeignKey(
         Employer, on_delete=models.CASCADE, blank=True, null=True)
+    apply_id = models.ForeignKey(
+        Application, on_delete=models.CASCADE, blank=True, null=True)
     int_link = models.CharField(
         max_length=300, default=None, blank=True, null=True)
     schedule_date = models.DateTimeField(default=timezone.now)
@@ -352,6 +360,7 @@ class Feedback(models.Model):
         max_length=500, default=None, blank=True, null=True)
     name = models.CharField(
         max_length=100, default=None, blank=True, null=True)
+    rating = models.IntegerField()
 
     class Meta:
         db_table = "feedback"
