+++ main/templates/footer.html
@@ -24,7 +24,9 @@
                                 <h3>For Candidates</h3>
                                 <ul class="pxp-footer-list">
                                     <li><a href="{% url 'main:jobs' %}">Find Jobs</a></li>
-                                    <li><a href="candidate-dashboard.html">Candidate Dashboard</a></li>
+                                    
+                                    <li><a href="{% url 'main:candidates' %}">Candidate Dashboard</a></li>
+
                                     <li><a href="#">My Applications</a></li>
                                     <li><a href="#">Favourite Jobs</a></li>
                                     <li><a href="#">My inbox</a></li>
