+++ main/templates/give_feedback.html
@@ -13,15 +13,47 @@
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/selectize.js/0.12.6/js/standalone/selectize.min.js" integrity="sha256-+C0A5Ilqmu4QcSPxrlGpaZxJ04VjsRjKu+G82kl5UJk=" crossorigin="anonymous"></script>
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/selectize.js/0.12.6/css/selectize.bootstrap3.min.css" integrity="sha256-ze/OEYGcFbPRmvCnrSeKbRTtjG4vGLHXgOqsyLFTRjg=" crossorigin="anonymous" />
-        
-
+        <link href="https://cdnjs.cloudflare.com/ajax/libs/mdb-ui-kit/6.4.1/mdb.min.css" rel="stylesheet"/>
         <link href="{% static 'css/bootstrap.min.css' %}" rel="stylesheet">
         <link rel="stylesheet" href="{% static 'css/font-awesome.min.css' %}">
         <link rel="stylesheet" href="{% static 'css/owl.carousel.min.css' %}">
         <link rel="stylesheet" href="{% static 'css/owl.theme.default.min.css' %}">
         <link rel="stylesheet" href="{% static 'css/animate.css' %}">
         <link rel="stylesheet" href="{% static 'css/style.css' %}">
-        <title>mpower - Contact us</title>
+        <title>mpower - Feedback</title>
+        <style>
+
+.star-rating {
+  
+  display:flex;
+  flex-direction: row-reverse;
+  font-size:2.0em;
+  justify-content:space-around;
+  padding:0 .2em;
+  text-align:center;
+
+  width:8em;
+  margin-bottom: 20px;
+}
+
+.star-rating input {
+  display:none;
+}
+
+.star-rating label {
+  color:#ccc;
+  cursor:pointer;
+}
+
+.star-rating :checked ~ label {
+  color:#f90;
+}
+
+.star-rating label:hover,
+.star-rating label:hover ~ label {
+  color:#fc0;
+}
+        </style>
     </head>
     <body>
         {% if messages %}
@@ -81,6 +113,27 @@
                                     <label for="contact-us-feeback" class="form-label">Feedback</label>
                                     <textarea type="text" class="form-control" id="contact-us-feedback" placeholder="Please provide feedback" name="feedback" required rows=20></textarea>
                                 </div>
+                                <div class="container">
+                                    <span id="rateMe1"></span>
+                                </div>
+                                <label for="rating" class="form-label">Overall Performance</label>
+
+
+<div class="star-rating">
+  <input type="radio" id="5-stars" name="rating" value="5" />
+  <label for="5-stars" class="star">&#9733;</label>
+  <input type="radio" id="4-stars" name="rating" value="4" />
+  <label for="4-stars" class="star">&#9733;</label>
+  <input type="radio" id="3-stars" name="rating" value="3" />
+  <label for="3-stars" class="star">&#9733;</label>
+  <input type="radio" id="2-stars" name="rating" value="2" />
+  <label for="2-stars" class="star">&#9733;</label>
+  <input type="radio" id="1-star" name="rating" value="1" />
+  <label for="1-star" class="star">&#9733;</label>
+</div>
+                                
+
+
                                 <button class="btn rounded-pill pxp-section-cta d-block">Send feedback</button>
                             </form>
                         </div>
@@ -97,5 +150,6 @@
         <script src="{% static 'js/nav.js' %}"></script>
         <script src="{% static 'js/main.js' %}"></script>
         <script src="//cdnjs.cloudflare.com/ajax/libs/moment.js/2.17.1/moment.min.js"></script>
+       
     </body>
 </html>
