+++ mpoweradmin/templates/mpoweradmin-index.html
@@ -44,8 +44,9 @@
 
         <div class="pxp-dashboard-side-panel d-none d-lg-block">
             <div class="pxp-logo">
-                <a href="{% url 'main:index' %}" class="pxp-animate"><span style="color: var(--pxpMainColor)">m</span>power</a>
+                <a href="{% url 'mpoweradmin:cdashboard' pk%}" class="pxp-animate"><span style="color: var(--pxpMainColor)">m</span>power</a>
             </div>
+            <!---->
 
             <nav class="mt-3 mt-lg-4 d-flex justify-content-between flex-column pb-100">
                 <div class="pxp-dashboard-side-label">admin tools</div>
