+++ mpoweradmin/views.py
@@ -600,192 +600,6 @@ def get_job(request,pk):
     job_data['phone'] = emp_.phone
     print("hello")
     return JsonResponse({'info': dumps(job_data, default=str)})
-class JobMatcher:
-    def __init__(self, jobseeker, job):
-        self.jobseeker = jobseeker
-        self.job = job
-    
-    
-
-    def calculate_match_percentage(self):
-        total_params = 5  
-        matched_params = 0
-
-        if self.jobseeker.skills and self.job.skills:
-            job_skills = set(self.job.skills.lower().split(','))
-            seeker_skills = set(self.jobseeker.skills.lower().split(','))
-            if job_skills.intersection(seeker_skills):
-                matched_params += 1
-
-        # Check if location matches
-        if self.job.location and self.jobseeker.location:
-            if self.job.location.lower() == self.jobseeker.location.lower():
-                matched_params += 1
-
-        # Check if notice period matches
-        if self.job.notice_period and self.jobseeker.notice_period:
-            if self.job.notice_period.lower() == self.jobseeker.notice_period.lower():
-                matched_params += 1
-
-        # Check if Experience matches
-        if self.jobseeker.experience and self.job.experience:
-            if int(self.jobseeker.experience) >= int(self.job.experience):
-                matched_params += 1
-
-        # Check if expected salary matches
-        if self.jobseeker.expsal and self.job.basicpay:
-            if self.jobseeker.expsal <= int(self.job.basicpay):
-                matched_params += 1
-
-        # Calculate match percentage
-        match_percentage = (matched_params / total_params) * 100
-
-        return match_percentage
-
-    def get_matching_details(self):
-        matching_details = {}
-
-        # Check if skills match
-        if self.jobseeker.skills and self.job.skills:
-            job_skills = set(self.job.skills.lower().split(','))
-            seeker_skills = set(self.jobseeker.skills.lower().split(','))
-            matching_details['skills'] = {
-                'match': list(job_skills.intersection(seeker_skills)),
-                'not_match': list(job_skills.difference(seeker_skills))
-            }
-            
-
-        # Check if location matches
-        if self.job.location and self.jobseeker.location:
-            matching_details['location'] = {
-                'match': True if self.job.location.lower() == self.jobseeker.location.lower() else False
-            }
-
-        # Check if notice period matches
-        if self.job.notice_period and self.jobseeker.notice_period:
-            matching_details['notice_period'] = {
-                'match': True if self.job.notice_period.lower() == self.jobseeker.notice_period.lower() else False
-            }
-
-        # Check if current salary matches
-        if self.jobseeker.cursal and self.job.basicpay:
-            matching_details['experience'] = {
-                'match': True if int(self.jobseeker.experience) >= int(self.job.experience) else False
-            }
-
-        # Check if expected salary matches
-        if self.jobseeker.expsal and self.job.basicpay:
-            matching_details['expected_salary'] = {
-                'match': True if self.jobseeker.expsal <= int(self.job.basicpay) else False
-            }
-
-        return matching_details
-
-def preprocess_text(text):
-    words = word_tokenize(text)
-
-    stop_words = set(stopwords.words('english'))
-    words = [word for word in words if word.lower() not in stop_words]
-
-    stemmer = PorterStemmer()
-    words = [stemmer.stem(word) for word in words]
-
-    preprocessed_text = ' '.join(words)
-    return preprocessed_text
-
-
-def extract_text_from_pdf(pdf_file):
-    pdf_text = ""
-    with open(pdf_file, 'rb') as file:
-        pdf_reader = PyPDF2.PdfReader(file)
-        num_pages = len(pdf_reader.pages)
-        for page_num in range(num_pages):
-            page = pdf_reader.pages[page_num]
-            pdf_text += page.extract_text()
-    return pdf_text
-def suggest(request, pk):
-    if request.method == "POST":
-        send = False
-        jobid = request.POST.get('jobid')
-        if (jobid):
-            try:
-                job_ = Jobs.objects.get(jobid=jobid)
-                if job_.status == 1 or job_.status == 3:
-                    send = True
-                matching_percentages = []
-                for job_seeker in JobSeeker.objects.all():
-                    important_data_jobs = f"{job_.title} {job_.jobdesc} {job_.fnarea} {job_.skills} {job_.experience} {job_.basicpay} {job_.location} {job_.industry} {job_.ugqual} {job_.pgqual} {job_.profile} {job_.jobtype} {job_.requirements} {job_.responsibilities} {job_.notice_period}"
-                    job_description = preprocess_text(important_data_jobs)
-
-                    important_data_jobseeker = f"{job_seeker.location} {job_seeker.experience} {job_seeker.skills} {job_seeker.basic_edu} {job_seeker.master_edu} {job_seeker.other_qual} {job_seeker.cursal} {job_seeker.expsal} {job_seeker.notice_period}"
-                    job_seeker_data = preprocess_text(important_data_jobseeker)
-                    resume_pdf_file = job_seeker.Resume.path
-                    resume_text = extract_text_from_pdf(resume_pdf_file)
-                    resume_text = preprocess_text(resume_text)
-                    total_job_seeker_data = f"{job_seeker_data}{resume_text}"
-                    vectorizer = CountVectorizer()
-
-                    job_description_vector = vectorizer.fit_transform([job_description])
-                    job_seeker_data_vector = vectorizer.transform([total_job_seeker_data])
-                    cosine_sim_job_seeker = cosine_similarity(job_description_vector, job_seeker_data_vector)
-
-                    job_matching_percentage = round(cosine_sim_job_seeker[0][0] * 100, 2)
-                    job_matcher = JobMatcher(jobseeker=job_seeker, job=job_)
-                    match_percentage = job_matcher.calculate_match_percentage()
-                    total_matching_percentage = round((job_matching_percentage + match_percentage) / 2, 2)
-                    matching_percentages.append((job_seeker.user_id, total_matching_percentage))
-                    matching_percentages.sort(key=lambda x: x[1], reverse=True)
-
-                top_matching_job_seekers = matching_percentages[:20]
-                if(send):
-                    email_list = []
-                    for i in top_matching_job_seekers:
-                        seek = JobSeeker.objects.get(user_id=i[0])
-                        email_list.append(seek.log_id.email)
-                    email_subject = "Job Alert"
-                    message = "You are among top 20 best profiles for this job"
-                    email = EmailMessage(email_subject, message, settings.EMAIL_HOST_USER, email_list)
-                    email.fail_silently = True
-                    # email.content_subtype = "html"
-                    email.send()
-                    
-                # print(top_matching_job_seekers)
-                serialized_list = json.dumps(top_matching_job_seekers)
-                job_.suggestions = serialized_list
-                job_.status = 4
-                job_.save()
-                return JsonResponse({'message': 'Candidate Suggestions are provided'})
-            except Jobs.DoesNotExist:
-                return JsonResponse({'message': 'Job not found'}, status=404)
-        else:
-            return JsonResponse({'message': 'Missing jobid'}, status=400)
-    else:
-        return JsonResponse({'message': 'Invalid request method'}, status=405)
-                # for i in top_matching_job_seekers:
-                #     seek = JobSeeker.objects.get(user_id=i[0])
-                #     single_candidate = {}
-                #     single_candidate['user_id'] = seek.user_id
-                #     single_candidate['name'] = seek.name
-                #     if seek.photo:
-                #         single_candidate['photo'] = seek.photo
-                #     single_candidate['score'] = i[1]
-                #     single_candidate['location'] = seek.location
-                #     candid.append(single_candidate)
-        
-    #     if jobid:
-    #         try:
-                
-    #             job = Jobs.objects.get(jobid=jobid)
-    #             job.status = 4  
-    #             job.save()
-    #             return JsonResponse({'message': 'Candidate Suggestions are provided'})
-    #         except Jobs.DoesNotExist:
-    #             return JsonResponse({'message': 'Job not found'}, status=404)
-    #     else:
-    #         return JsonResponse({'message': 'Missing jobid'}, status=400)
-    # else:
-    #     return JsonResponse({'message': 'Invalid request method'}, status=405)
-
 
 def subscriptions(request, pk):
     return render(request, 'subscriptions-mpoweradmin.html', {'pk': pk})
